Extended sheading functions to support negative sheadin factors; Refactored interpolation functions to support negative numbers also

+++ lib/auxiliaries/pixel_interpolation.go
@@ -13,10 +13,10 @@ type Pixel struct {
     segment delimited by the  2 pixels.
 */
 func Pixel_linear_interpolation(px1, px2 Pixel, x float64) (Pixel) {
-    r_aux := interp.Linear_interpolation(px1.R, px2.R, x)
-    g_aux := interp.Linear_interpolation(px1.G, px2.G, x)
-    b_aux := interp.Linear_interpolation(px1.B, px2.B, x)
-    a_aux := interp.Linear_interpolation(px1.A, px2.A, x)
+    r_aux := uint32(interp.Linear_interpolation(int32(px1.R), int32(px2.R), x))
+    g_aux := uint32(interp.Linear_interpolation(int32(px1.G), int32(px2.G), x))
+    b_aux := uint32(interp.Linear_interpolation(int32(px1.B), int32(px2.B), x))
+    a_aux := uint32(interp.Linear_interpolation(int32(px1.A), int32(px2.A), x))
 
     return Pixel{r_aux, g_aux, b_aux, a_aux}
 }
@@ -26,10 +26,10 @@ func Pixel_linear_interpolation(px1, px2 Pixel, x float64) (Pixel) {
     square delimited by the 2 pixels.
 */
 func Pixel_bilinear_interpolation(px11, px12, px21, px22 Pixel, x, y float64) (Pixel) {
-    r_aux := interp.Bilinear_interpolation(px11.R, px12.R, px21.R, px22.R, x, y)
-    g_aux := interp.Bilinear_interpolation(px11.G, px12.G, px21.G, px22.G, x, y)
-    b_aux := interp.Bilinear_interpolation(px11.B, px12.B, px21.B, px22.B, x, y)
-    a_aux := interp.Bilinear_interpolation(px11.A, px12.A, px21.A, px22.A, x, y)
+    r_aux := uint32(interp.Bilinear_interpolation(int32(px11.R), int32(px12.R), int32(px21.R), int32(px22.R), x, y))
+    g_aux := uint32(interp.Bilinear_interpolation(int32(px11.G), int32(px12.G), int32(px21.G), int32(px22.G), x, y))
+    b_aux := uint32(interp.Bilinear_interpolation(int32(px11.B), int32(px12.B), int32(px21.B), int32(px22.B), x, y))
+    a_aux := uint32(interp.Bilinear_interpolation(int32(px11.A), int32(px12.A), int32(px21.A), int32(px22.A), x, y))
 
 return Pixel{r_aux, g_aux, b_aux, a_aux}
 }
